Added login logout registration functionality.
Added login logout registration functionality with firebase.
