+++ assignmentB/frontend/package.json
@@ -11,12 +11,14 @@
   },
   "dependencies": {
     "axios": "^1.7.7",
+    "firebase": "^10.14.1",
     "framer-motion": "^11.5.4",
     "js-cookie": "^3.0.5",
     "next": "^14.2.5",
     "next-auth": "^4.24.7",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
+    "react-firebase-hooks": "^5.1.1",
     "react-hook-form": "^7.52.2",
     "react-icons": "^5.3.0",
     "sass": "^1.77.8"
