+++ assignmentB/frontend/src/components/LoginPage.js
@@ -1,0 +1,46 @@
+// src/components/LoginPage.js
+import React, { useState } from 'react';
+import { signInWithEmailAndPassword } from 'firebase/auth';
+import { useNavigate } from 'react-router-dom';
+import { auth } from '../firebase';
+
+const LoginPage = () => {
+  const [email, setEmail] = useState('');
+  const [password, setPassword] = useState('');
+  const [error, setError] = useState('');
+  const navigate = useNavigate();
+
+  const handleLogin = async (e) => {
+    e.preventDefault();
+    try {
+      await signInWithEmailAndPassword(auth, email, password);
+      navigate('/admin-dashboard'); // Redirect to the Admin dashboard
+    } catch (err) {
+      setError('Failed to log in. Please check your credentials.');
+    }
+  };
+
+  return (
+    <div className="login-container">
+      <h2>Admin Login</h2>
+      {error && <p>{error}</p>}
+      <form onSubmit={handleLogin}>
+        <input
+          type="email"
+          placeholder="Email"
+          value={email}
+          onChange={(e) => setEmail(e.target.value)}
+        />
+        <input
+          type="password"
+          placeholder="Password"
+          value={password}
+          onChange={(e) => setPassword(e.target.value)}
+        />
+        <button type="submit">Log In</button>
+      </form>
+    </div>
+  );
+};
+
+export default LoginPage;
