+++ assignmentB/frontend/src/components/PopulateNavBar.tsx
@@ -6,67 +6,49 @@ import NotificationBadge from "../components/notification/NotificationBadge"; //
 import { useEffect, useState } from "react";
 import axios from "axios";
 import React from 'react';
-
+import { useAuthState } from 'react-firebase-hooks/auth';
+import { auth } from '../firebase'; // Make sure you adjust this path to your firebase setup
+import { signOut } from 'firebase/auth';
+import { useRouter } from 'next/router';
 
 const PopulatedNavBar = () => {
-  const [pendingCount, setPendingCount] = useState(0);
-  const [moderatedCount, setModeratedCount] = useState(0);
-  const [moderatedAndRejectedCount, setModeratedAndRejectedCount] = useState(0);
-
+  const [pendingCount, setPendingCount] = useState<number | null>(null); // null while loading
+  const [moderatedCount, setModeratedCount] = useState<number | null>(null);
+  const [moderatedAndRejectedCount, setModeratedAndRejectedCount] = useState<number | null>(null);
+  const [user] = useAuthState(auth); // Get the current authenticated user
+  const router = useRouter();
 
   useEffect(() => {
-    // Fetch the number of pending moderation articles
-    const fetchPendingCount = async () => {
+    const fetchData = async () => {
       try {
-        const response = await axios.get('http://localhost:3001/api/moderation/pending-count');
-        console.log("Full API response: ", response);
-        setPendingCount(response.data); // Adjust based on the API response
-        console.log("pending count checking: ", response.data);
+        const [pendingResponse, moderatedResponse, rejectedResponse] = await Promise.all([
+          axios.get('http://localhost:3001/api/moderation/pending-count'),
+          axios.get('http://localhost:3001/api/analysis'),
+          axios.get('http://localhost:3001/api/submitter/moderated-rejected-count')
+        ]);
+
+        setPendingCount(pendingResponse.data);
+        setModeratedCount(moderatedResponse.data.length); // Assuming it's an array
+        setModeratedAndRejectedCount(rejectedResponse.data);
+
       } catch (err) {
-        console.error("Error fetching pending moderation count", err);
+        console.error("Error fetching notification data", err);
       }
     };
 
-    const fetchModeratedAndRejectedCount  = async () => {
-      try {
-        const response = await axios.get('http://localhost:3001/api/submitter/moderated-rejected-count');
-        setModeratedAndRejectedCount(response.data); 
-        console.log("rejected count checking: ", response.data);
-      } catch (err) {
-        console.error("Error fetching rejected article count", err);
-      }
-    };
+    fetchData();
 
-
-    const fetchModeratedCount = async () => {
-      try {
-        const response = await axios.get('http://localhost:3001/api/analysis');
-        console.log("Full API response 22222: ", response);
-        setModeratedCount(response.data.length);
-      } catch (err) {
-        console.log("Error fetching moderated count from the analysis api", err);
-      }
-    }
-
-    // Initial fetch
-    fetchPendingCount();
-    fetchModeratedCount();
-    fetchModeratedAndRejectedCount();
-
-    // Poll every 3 seconds
-    const pendingIntervalId = setInterval(fetchPendingCount, 1000);
-    const moderatedIntervalId = setInterval(fetchModeratedCount, 1000);
-    const moderatedAndRejectedId = setInterval(fetchModeratedAndRejectedCount, 3000);
+    // Poll every 5 seconds to reduce the load
+    const intervalId = setInterval(fetchData, 5000);
 
     // Cleanup interval on component unmount
-    return () => {
-      clearInterval(pendingIntervalId);
-      clearInterval(moderatedIntervalId);
-      clearInterval(moderatedAndRejectedId);
-
-    };
-
+    return () => clearInterval(intervalId);
   }, []);
+
+  const handleLogout = async () => {
+    await signOut(auth); // Sign out the user
+    router.push('/login'); // Redirect to the login page after logout
+  };
 
   return (
     <NavBar>
@@ -84,29 +66,47 @@ const PopulatedNavBar = () => {
         </NavDropdown>
       </NavItem>
 
-      {/* New dropdown for admin/moderation related pages */}
-      <NavItem dropdown route="/admin">
-        Admin <IoMdArrowDropdown />
-        <NavDropdown>
-          <NavItem route="/moderation">
-            Moderation
-            {/* Display the red dot and count next to Moderation */}
-            <NotificationBadge count={pendingCount} />
-            </NavItem>
-          <NavItem route="/analysis">
-            Analysis
-            {/* Display the red dot and waiting for analysis count next to Analysis */}
-            <NotificationBadge count={moderatedCount} />
-            </NavItem>
-            <NavItem route="/submitter">
-            Submitter Dashboard
-            {/* show the count of notifications */}
-            <NotificationBadge count={moderatedAndRejectedCount} />
-            </NavItem>
-          <NavItem route="/admin">Admin Dashboard</NavItem>
-          <NavItem route="/search">Search</NavItem>
-        </NavDropdown>
-      </NavItem>
+      {/* Conditional rendering for Admin-related links if user is authenticated */}
+      {user ? (
+        <>
+          <NavItem dropdown route="/admin">
+            Admin <IoMdArrowDropdown />
+            <NavDropdown>
+              <NavItem route="/moderation">
+                Moderation
+                {pendingCount !== null ? (
+                  <NotificationBadge count={pendingCount} />
+                ) : (
+                  <span>Loading...</span>
+                )}
+              </NavItem>
+              <NavItem route="/analysis">
+                Analysis
+                {moderatedCount !== null ? (
+                  <NotificationBadge count={moderatedCount} />
+                ) : (
+                  <span>Loading...</span>
+                )}
+              </NavItem>
+              <NavItem route="/submitter">
+                Submitter Dashboard
+                {moderatedAndRejectedCount !== null ? (
+                  <NotificationBadge count={moderatedAndRejectedCount} />
+                ) : (
+                  <span>Loading...</span>
+                )}
+              </NavItem>
+              <NavItem route="/admin">Admin Dashboard</NavItem>
+              <NavItem route="/search">Search</NavItem>
+            </NavDropdown>
+          </NavItem>
+          <button onClick={handleLogout} className="bg-red-500 text-white px-4 py-2 rounded">
+            Logout
+          </button>
+        </>
+      ) : (
+        <NavItem route="/login">Login</NavItem>
+      )}
     </NavBar>
   );
 };
